backend/requirements-jetson.txt
absl-py==0.7.0
acoular==18.1
astor==0.7.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cycler==0.10.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-SocketIO==3.2.1
gast==0.2.2
greenlet==0.4.15
grpcio==1.18.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
llvmlite==0.27.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
monotonic==1.5
multiprocessing-logging==0.2.6
numba==0.42.0
numexpr==2.6.9
numpy==1.16.1
pkg-resources==0.0.0
protobuf==3.6.1
pycparser==2.19
pyface==6.0.0
pyparsing==2.3.1
python-dateutil==2.7.5
python-engineio==3.3.0
python-socketio==3.1.2
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sounddevice==0.3.12
tables==3.4.4
tensorboard==1.9.0
tensorflow-gpu==1.9.0+nv18.8
termcolor==1.1.0
traits==5.0.0
traitsui==6.0.0
urllib3==1.24.1
Werkzeug==0.14.1
